+++ core/object_node.go
@@ -1,5 +1,7 @@
 package core
 
+// ObjectNodeInterface describes object node, most common node of game engine,
+// which used to render primitives or textures, or to group multiple nodes.
 type ObjectNodeInterface interface {
 	BaseNodeInterface
 
@@ -11,6 +13,8 @@ type ObjectNodeInterface interface {
 	SetTexture(texture *Texture)
 }
 
+// ObjectNode is an implementation of ObjectNodeInterface
+// ObjectNode have to be initialized with NewObjectNode()
 type ObjectNode struct {
 	BaseNode
 
@@ -26,16 +30,22 @@ func NewObjectNode() *ObjectNode {
 	}
 }
 
+// GetNodeType used internally to determine which node passed to polymorph function
 func (n *ObjectNode) GetNodeType() NodeType { return ObjectNodeType }
+
+// GetChildren returns slice of all children attached to this node
 func (n *ObjectNode) GetChildren() []BaseNodeInterface {
 	return n.children
 }
 
+// AddChild adds child to this node
 func (n *ObjectNode) AddChild(child BaseNodeInterface) {
 	n.children = append(n.children, child)
 	child.SetParent(n)
 }
 
+// RemoveChild tries to find and remove child based on pointer equality,
+// returns true in success and false if child was not found
 func (n *ObjectNode) RemoveChild(child BaseNodeInterface) bool {
 	newNodes := make([]BaseNodeInterface, 0, len(n.children))
 	found := false
@@ -44,6 +54,7 @@ func (n *ObjectNode) RemoveChild(child BaseNodeInterface) bool {
 			newNodes = append(newNodes, c)
 		} else {
 			found = true
+			c.SetParent(nil)
 		}
 	}
 
@@ -54,10 +65,12 @@ func (n *ObjectNode) RemoveChild(child BaseNodeInterface) bool {
 	return found
 }
 
+// GetTexture returns set texture or nil if no texture wa provided to node
 func (n *ObjectNode) GetTexture() *Texture {
 	return n.texture
 }
 
+// SetTexture sets texture of the node
 func (n *ObjectNode) SetTexture(texture *Texture) {
 	n.texture = texture
 }