+++ primitive/color.go
@@ -2,4 +2,5 @@ package primitive
 
 import "github.com/veandco/go-sdl2/sdl"
 
+// Color is a type alias to use inside game engine
 type Color = sdl.Color