+++ resource/image.go
@@ -6,6 +6,8 @@ import (
 	"github.com/veandco/go-sdl2/sdl"
 )
 
+// Image used to link path to file on disk and its loaded content,
+// have to be initialized with NewFont
 type Image struct {
 	path string
 
@@ -21,6 +23,8 @@ func NewImage(path string) *Image {
 	}
 }
 
+// Reload load content of file and stores it as pixel data in memory.
+// Reload called automatically if it was not called before.
 func (i *Image) Reload() {
 	surf, err := img.Load(i.path)
 	if err != nil {
@@ -31,6 +35,7 @@ func (i *Image) Reload() {
 	}
 }
 
+// GetSurface is an internal function, returns image representation used to render it.
 func (i *Image) GetSurface() *sdl.Surface {
 	if !i.loaded {
 		i.Reload()